+++ src/frontend/ast.mli
@@ -5,7 +5,13 @@ type location = private {
   defined_at : Common.Span.t option;
 }
 
-type declaration_modifier_kind = Public | Static
+type declaration_modifier_kind =
+  | Public
+  | Static
+  | Interrupt
+  | Has_error_code
+  | Argument_pop
+  | No_argument_pop
 
 type declaration_modifier = private {
   kind : declaration_modifier_kind;